Implementation of MissingCouplingRule
Implemented above mentioned rule.
In base class rule, the check method needs *args to allow child classes to pass the graph.
It might be better to just directly pass graph, as I do not see an option where we do not use the graph (but something else) for checking the rule.
In cli.py an optional cli input parameter for severity was added in order to print rules based on severity